custom error implementation to thiserror
