+++ Cargo.toml
@@ -26,8 +26,7 @@ enum-iterator = "0"
 getset = "0"
 git2 = { version = "0", optional = true, default-features = false }
 rustc_version = { version = "0.3.1", optional = true }
-serde = "1"
-serde_derive = "1"
+thiserror = "1"
 
 [build-dependencies]
 chrono = "0"
